# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CortaAi API",
    "description": "API completa do sistema CortaAi — agendamentos de barbearia.\n\n**Autenticação:** Firebase Bearer Token.\n\nVariáveis de ambiente:\n- `base_url`: `https://api.cortaai.shop`\n\n**Fluxo básico:**\n1. `POST /api/auth/email/login` → salvar `idToken`\n2. Usar `Authorization: Bearer {idToken}` nas rotas protegidas",
    "version": "1.0.0",
    "contact": {
      "name": "CortaAi",
      "email": "contato@cortaai.shop"
    }
  },
  "servers": [
    {
      "url": "https://api.cortaai.shop",
      "description": "Produção"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local (Docker)"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Firebase ID Token"
      }
    },
    "schemas": {
      "EmailRegisterRequest": {
        "type": "object",
        "required": [
          "email",
          "password",
          "userType"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "usuario@teste.com"
          },
          "password": {
            "type": "string",
            "example": "Senha@123"
          },
          "userType": {
            "type": "string",
            "enum": [
              "CUSTOMER",
              "BARBER"
            ],
            "example": "CUSTOMER"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "usuario@teste.com"
          },
          "password": {
            "type": "string",
            "example": "Senha@123"
          }
        }
      },
      "ForgotPasswordRequest": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "usuario@teste.com"
          }
        }
      },
      "ChangePasswordRequest": {
        "type": "object",
        "required": [
          "idToken",
          "newPassword"
        ],
        "properties": {
          "idToken": {
            "type": "string",
            "description": "Firebase idToken da sessão atual (obtido no login)",
            "example": "eyJhbGciOiJSUzI1NiJ9..."
          },
          "newPassword": {
            "type": "string",
            "description": "Nova senha (mínimo 6 caracteres)",
            "minLength": 6,
            "example": "NovaSenha@123"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "idToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "expiresIn": {
            "type": "string"
          },
          "localId": {
            "type": "string"
          },
          "profile": {
            "$ref": "#/components/schemas/AuthResponse"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "userType": {
            "type": "string",
            "enum": [
              "CUSTOMER",
              "BARBER"
            ]
          },
          "barbershopId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "isOwner": {
            "type": "boolean"
          }
        }
      },
      "VerifyTokenRequest": {
        "type": "object",
        "required": [
          "idToken",
          "userType"
        ],
        "properties": {
          "idToken": {
            "type": "string"
          },
          "userType": {
            "type": "string",
            "enum": [
              "CUSTOMER",
              "BARBER"
            ]
          }
        }
      },
...
      "JoinRequest": {
        "type": "object",
        "required": [
          "cnpj"
        ],
        "properties": {
          "cnpj": {
            "type": "string",
            "example": "12345678000199"
          }
        }
      },
...
      "EmailExistsResponse": {
        "type": "object",
        "properties": {
          "exists": {
            "type": "boolean",
            "example": true
          },
          "userType": {
            "type": "string",
            "enum": ["CUSTOMER", "BARBER"],
            "nullable": true,
            "example": "CUSTOMER",
            "description": "Tipo de usuário encontrado, null se não existir"
          }
        }
      },
...
    }
  },
...
  "paths": {
    "/api/auth/email/register": {
      "post": {
        "tags": [
          "01 - Auth"
        ],
        "summary": "Registrar usuário (email/senha)",
        "operationId": "registerEmail",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailRegisterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário registrado"
          },
          "400": {
            "description": "Dados inválidos"
          }
        }
      }
    },
    "/api/auth/email/login": {
      "post": {
        "tags": [
          "01 - Auth"
        ],
        "summary": "Login (email/senha)",
        "operationId": "loginEmail",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token retornado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas"
          }
        }
      }
    },
    "/api/auth/email/forgot-password": {
      "post": {
        "tags": [
          "01 - Auth"
        ],
        "summary": "Recuperação de senha",
        "description": "Envia um e-mail de redefinição de senha via Firebase. O Firebase dispara o e-mail automaticamente com um link seguro. Rota pública.",
        "operationId": "forgotPassword",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "E-mail de recuperação enviado com sucesso"
          },
          "400": {
            "description": "E-mail inválido ou ausente"
          },
          "401": {
            "description": "E-mail não encontrado no Firebase"
          }
        }
      }
    },
    "/api/auth/email/change-password": {
      "post": {
        "tags": [
          "01 - Auth"
        ],
        "summary": "Alterar senha (usuário autenticado)",
        "description": "Altera a senha do usuário autenticado. Requer o idToken da sessão atual (obtido no login). Após a alteração, o Firebase invalida todas as sessões anteriores. Rota pública — o idToken vai no corpo.",
        "operationId": "changePassword",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Senha alterada com sucesso"
          },
          "400": {
            "description": "Dados inválidos ou senha muito curta (mín. 6 caracteres)"
          },
          "401": {
            "description": "Token expirado ou inválido"
          }
        }
      }
    },
    "/api/auth/verify": {
      "post": {
        "tags": [
          "01 - Auth"
        ],
        "summary": "Verificar e registrar token Firebase",
        "operationId": "verifyToken",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyTokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token verificado"
          }
        }
      }
    },
...
    "/api/barbershops/join-request": {
      "post": {
        "tags": [
          "04 - Barbershops"
        ],
        "summary": "Solicitar entrada em barbearia (barbeiro)",
        "operationId": "joinRequest",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Solicitação enviada"
          }
        }
      }
    },
...
    "/api/auth/email/exists": {
      "get": {
        "tags": [
          "01 - Autenticação"
        ],
        "summary": "Verificar se e-mail já está cadastrado",
        "description": "Endpoint público. Retorna se o e-mail existe na base e o tipo de usuário (CUSTOMER ou BARBER).",
        "operationId": "emailExists",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "email"
            },
            "example": "usuario@teste.com"
          }
        ],
        "responses": {
          "200": {
            "description": "Resultado da verificação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailExistsResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}